client-53 Написать скрипт зависимости цен переходов действий ии от ха…

+++ Assets/AI/Scripts/Entity/ActionCost.cs
@@ -7,7 +7,7 @@ namespace AI.Scripts.Entity
     public class ActionCost 
     {
         [SerializeField]
-        public int Cost;
+        public float Cost;
         [SerializeField]
         public GameObject Action;
         [SerializeField]
